backend/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
flask-cors==4.0.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
psycopg2-binary

